requirements.txt
django
djangorestframework
mssql-django
django-cors-headers